requirements.txt
streamlit
validators
python-dotenv
yt-dlp
openai-whisper
torch
langchain
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
unstructured
